Simplify Animation logic.
Push repeat field down to different animation types and separate animation Performers to repeatable and non-repeatable.

This allows animations that run on parallel to have different repetition patterns and allow for a composite animation to have clearer finish semantics. A composite animation finishes if any part of it finishes.

Also change the return logic of Animator::Progress() to match that of Performer::Progress(); returns true when animation is finished.