Add support for instant repeatable performers
Add support for a new animation type; repeatable animations that are applied instantly.

Also fix a bug in an odd case for repeatable animations where repeat == 0 && delay == 0. This would cause an infinite loop. Now it only applies the animation once per call regardless of time passed since last call. This is a cornercase that should not be used in scripts, but the engine doesn't malfunction.